+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.5
 - **Attach to…**: Add `tracciatto.socketSearchRoot` to set search directory for `rdbg --util=list-socks`.
-- `tracciatto.patchMaxInspectedValueLength` accepts any value greater than 0, instead of foricing more than 180, for consistent UX.
+- `tracciatto.patchMaxInspectedValueLength` accepts any value greater than 0, instead of forcing more than 180, for consistent UX.
 
 ## 0.3.4
 - Fix: **Set Value** cannot update the inner fields of values created or assigned through **Set Value** (due to missing variable‑reference tracking).